collection_to_csv/requirements.txt
greenlet==3.0.3
psycopg2==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.30
typing_extensions==4.12.2
